Move hex encode/decode to a separate header since including io-util on Windows causes a compilation failure
